requirements.txt
llama-parse
torch
torch-vision
python-dotenv

unstructured
sentence-transformers
llama-index-embeddings-huggingface